Move lockfree bucket to ydb/library
