+++ ydb/library/ya.make
@@ -14,6 +14,7 @@ RECURSE(
     grpc
     http_proxy
     keys
+    lockfree_bucket
     logger
     login
     mkql_proto
